refactor: isolate cli runtime options
