+++ src/cli/runtime-options/byte-format.js
@@ -1,6 +1,6 @@
 import { Buffer } from "node:buffer";
 import { isFiniteNumber } from "../../shared/number-utils.js";
-import { coercePositiveInteger } from "./dependencies.js";
+import { coercePositiveInteger } from "../shared/dependencies.js";
 import { createIntegerOptionToolkit } from "../core/integer-option-toolkit.js";
 
 const BYTE_UNITS = Object.freeze(["B", "KB", "MB", "GB", "TB", "PB"]);
