+++ src/cli/runtime-options/sample-limit-toolkit.js
@@ -1,0 +1,47 @@
+import { coerceNonNegativeInteger } from "../shared/dependencies.js";
+import { createIntegerOptionToolkit } from "../core/integer-option-toolkit.js";
+
+/**
+ * Create a CLI sample limit toolkit that mirrors the previous standalone
+ * modules while centralizing the error messaging and coercion boilerplate.
+ *
+ * @param {object} parameters
+ * @param {number} parameters.defaultValue Baseline limit before overrides.
+ * @param {string} [parameters.envVar] Environment variable that overrides the
+ *        default when defined.
+ * @param {string} parameters.subjectLabel Descriptive label used in error
+ *        messages (e.g. "Skipped directory").
+ * @param {string} [parameters.defaultValueOption="defaultLimit"] Alias passed
+ *        to the resolver so commands can customise option names.
+ * @returns {{
+ *   getDefault: () => number | undefined,
+ *   setDefault: (value: unknown) => number | undefined,
+ *   resolve: (value: unknown, options?: Record<string, unknown>) =>
+ *       number | null | undefined,
+ *   applyEnvOverride: (env?: NodeJS.ProcessEnv | null | undefined) =>
+ *       number | undefined
+ * }}
+ */
+export function createSampleLimitToolkit({
+    defaultValue,
+    envVar,
+    subjectLabel,
+    defaultValueOption = "defaultLimit"
+} = {}) {
+    const normalizedLabel = subjectLabel ?? "Sample";
+
+    const createSampleLimitErrorMessage = (received) =>
+        `${normalizedLabel} sample limit must be a non-negative integer (received ${received}). Provide 0 to suppress the sample list.`;
+
+    const createSampleLimitTypeErrorMessage = (type) =>
+        `${normalizedLabel} sample limit must be provided as a number (received type '${type}').`;
+
+    return createIntegerOptionToolkit({
+        defaultValue,
+        envVar,
+        baseCoerce: coerceNonNegativeInteger,
+        createErrorMessage: createSampleLimitErrorMessage,
+        typeErrorMessage: createSampleLimitTypeErrorMessage,
+        defaultValueOption
+    });
+}
